creait/oliveyoung_scraper/pyproject.toml
[project]
name = "creait"
version = "0.1.0"
description = "Instagram Product Trust Analysis System"
readme = {content-type = "text/markdown", file = "README.md", charset = "utf-8"}
requires-python = ">=3.13"
dependencies = [
    # AI/LLM
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    # 웹 스크래핑
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    # 데이터베이스
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # 데이터 검증
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # 유틸리티
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "pandas>=2.3.3",
    "playwright-stealth>=2.0.0",
    "undetected-chromedriver>=3.5.5",
    "setuptools>=80.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
